refactor: split hardening hotspots by responsibility
